Record architecture decisions
We need to record the architectural decisions made on this project.

We will use Architecture Decision Records, as described by Michael
Nygard in this article:
http://thinkrelevance.com/blog/2011/11/15/documenting-architecture-decis
ions

This adds a decision documenting our use of Touchstone SAML for
authentication and Heroku Pipelines for deployment.

Also documents current ENV and adds warning about dev auth.