Add roles
Users can have four roles:
* basic (i.e. thesis depositors)
* processor
* thesis administrator (access to admin dashboards for add/edit)
* sysadmin

This adds those roles; updates existing tests to pass; and adds tests
of the new roles.

It also uncovered a fun bug: we couldn't actually create theses through
the admin dashboards, because we asked for grad_date (the column in the
model), but our post data then got rejected because the before_create
steps of validating graduation month and year failed. Rather than
rewriting the model logic to account for both cases, I made the admin
form consistent with the behavior described in the model.