Remove broken migration and rebuild schema
This migration was needed on localhost because my database had ended up
in a weird state not properly reflected by db/schema.rb due to an
earlier iteration of the name-adding work. Staging had gone through
the same history, because it was necessary to deploy the
branch-in-progress to staging in order to test Touchstone integration.
However, production had NOT gone through the same database history,
so it did not need the name removal migration, and the attempt to
deploy the earlier version of the branch failed.

In this work, I have forcibly dropped and recreated (not just reset)
local and staging databases and reverted the merge to master. This will
let me recreate the database consistent with the prior version of
master and test that we can roll forward from it.