Updates pr build pattern match to allow fake auth
The PR builds now have incremented over 99 which seems to have
triggered a pattern match problem as they removed an extra character
from the middle of the expected URL to allow for the extra character
in the number to fit in their url size.

This should keep us safe for another 900 builds or so.