Bugfix out-of-order dates
Chartkick renders the data series in the order it receives them, meaning
that the X axis can end up out of order if there are some dates
that occur only in later series. We create a dummy series, with all
possible dates but no data, in order to force Chartkick to render the
entire X axis before rendering data atop it.