Bugfix incorrect dropdowns for date filter
Date filtering causes a status parameter to be applied to the URL.
When this param is empty, by_status('') returns only active theses, so
the earliest_year and latest_year numbers are wrong - they don't match
the actual queryset on the page.

Maybe by_status has the wrong behavior, but that can be a different
ticket.