Allow processors to view all Theses
https://mitlibraries.atlassian.net/browse/TI-110

This bug has been around a while, but was only noticeable to staff that are thesis_admins or processors but also _not_ general admins.

The presence of effectively two roles systems (the named roles and the "admin" role) is technical debt and a new ticket will be opened to address that. The new ticket is https://mitlibraries.atlassian.net/browse/TI-115