Redirect back to filtered path when updating records
When working from the theses process list, marking a thesis complete, withdrawn or adding a note returned the processor to the process_path. However, that didn't take into account any filters or sorting they had done.

By using the redirect_back with a fallback to the process_path we should safely be able to return them to the same filtered list they started from while returning to the unfiltered list if the request doesn't have that information for some reason.

https://mitlibraries.atlassian.net/browse/TI-116