Update dependencies and change default branch
Why these changes are being introduced:

Several dependencies have updates available, and we haven't yet
changed this repo's default branch to main.

Relevant ticket(s):

N/A

How this addresses that need:

This changes the default branch to main in GitHub Actions and
updates the following dependencies:
* aws-sdk-s3: 1.83.1 → 1.86.0 (minor)  depfu
* puma: 5.0.4 → 5.1.0 (minor)  depfu
* listen: 3.2.1 → 3.3.3 (minor)  depfu
* simplecov: 0.19.1 → 0.20.0 (major)  depfu
* capybara: 3.33.0 → 3.34.0 (minor)  depfu
* rubocop: 1.0.0 → 1.4.2 (minor)  depfu
* bootsnap: 1.4.9 → 1.5.1 (minor)  depfu

Side effects of this change:

Just the residual effects of changing the main branch, which have
been addressed.