Update PR template
Why these changes are being introduced:

This repo uses an older version of our PR template, which has been
superseded.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETS-83

How this addresses that need:

This adds the new PR template that we intend to use for EngX repos.

Side effects of this change:

This assumes that all committers maintain a .gitmessage file locally
with the following (or similar) headings as in this commit message.