Fix generated dashboard UI for viewing transfers
Why these changes are being introduced:

* The UI generated by the administrate gem for viewing transfers is broken
due to missing templates for file attachments and blobs.
* The gem doesn't support ActiveStorage, so we need to define custom
fields as described here: https://github.com/thoughtbot/administrate/blob/master/docs/adding_custom_field_types.md

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETS-88

How this addresses that need:

* Defines custom fields for attachments.
* Removes blob as an attribute in all transfer templates to reduce
redundancy.
* Removes attachments and blobs as form attributes, since that doesn't
appear to work anyway.
* Changes grad date format in show page to %Y %B.
* Renders grad date and transfer date in index page.

Side effects of this change:

I chose not to define a custom field for blobs, as I couldn't think of a
scenario in which we would want separate fields for blobs and attachments
in the UI. However, this might change if we want to allow users to create
and edit full transfers through the admin dashboard.