Update thesis model and corresponding views
Why these changes are being introduced:

The thesis model needs to incorporate changes in v1.2 of the data model. Closes ETS-90 and ETS-104.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETS-90
* https://mitlibraries.atlassian.net/browse/ETS-104

How this addresses that need:

* Makes title and abstract columns nullable.
* Adds author_note, files_complete, metadata_complete, and
publication_status columns.
* Renames note column to processor_note.
* Adds the above fields to the thesis admin dashboard show and form views.
* Changes instances of 'note' in views to 'processor note'.

Side effects of this change:

* We will need to figure out where (outside of the admin dashboard) we
want these fields to render, and which users should be able to edit them.
* We will need to update the annotate method in the controller to allow
for annotating author notes.
* The status column no longer exists in v1.2, but removing it is beyond
the scope of this ticket.