Adds fields to active_storage_attachment
Why are these changes being introduced:

* Our workflows require the ability to track the reason a file is being
  submitted (primary thesis file, supplementary files, signature page,
  etc)

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-30

How does this address that need:

* By extending the active_storage_attachment class and adding two fields
  to the associated database table, we can easily provide the info our
  app needs
* The extra module we load into the class is only necessary to allow an
  enum, but it has a nice side effect of allowing us to include the
  module with our models to be more visible to developers that there is
  something non-default in our active_storage_attachment class.

Document any side effects to this change:

By extending the active_storage_attachment class, it is possible future
changes to ActiveStorage may be in conflict with our changes. This feels
quite unlikely, but as it's possible it feels worth noting.

We are using the terminology "purpose" for the "type" of file. The
reason being that a file "type" is commonly used to describe things like
"it's a PDF" or "it's a zip file". As we _also_ care about that meaning
of "type" in this application and already store that by default in
ActiveStorage it felt useful longer term to choose another label to
represent the "purpose" of the file being submitted.

I've normalized to a single way of accessing attachments in
administrate. There should be no negative side effects, but this change
was out of scope for this work but it felt like useful maintenance while
I was working in this area.