Update User model to match v1.4 of data model
Why these changes are being introduced:

The latest version of the data model includes changes to the User table
that are not currently reflected in the codebase.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-76

How this addresses that need:

* Adds kerberos_id, orcid, display_name, middle_name, preferred_name fields
* Changes name method to use preferred_name if available
* Removes email from method and creates separate name method for
processing queue (where we still want names)
* Adds lifecycle callbacks to handle validation errors and update name
whenever user records are updated
* Updates user dashboard to use display_name instead of given_name and
surname in show and index pages
* Adds preferred_name, kerberos_id and orcid to show page of user dashboard
* Adds preferred_name and orcid to edit page of user dashboard
* Adds orcid to new page of user dashboard
* Updates display_resource method in user dashboard to use preferred_name

Side effects of this change:

* There are some assumptions made about names here that we may need to
adjust later, pending user feedback (though, most of my concerns are
mitigated by our allowing any form for preferred names).
* The model changes will cause problems for pre-existing records, but
this shouldn't be an issue since we're not migrating data.
* If not present, kerb is generated from uid, and uid from kerb. This is
necessary for validation purposes, but it assumes consistent and
accurate data from users and Touchstone.