Skip test that is failing in GitHub Actions
Why these changes are being introduced:

One of our integration tests is failing in GitHub Actions (but not
locally), which is breaking all of our builds.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-173

How this addresses that need:

This skips the problem test.

Side effects of this change:

The ability to edit department_thesis through the admin dashboard will
not be tested until we fix the root cause.