Add :restrict_with_error option to Thesis-Hold association
Why these changes are being introduced:

When a user tries to delete a thesis with a hold in the admin UI,
the action fails due to the FK constraint. This is the behavior we want,
but we'd like to swallow the ActiveRecord::InvalidForeignKey error so
processors don't see a Rails error page.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-153

How this addresses that need:

This adds the :restrict_with_error option to the has_many: :holds
association on the Thesis model, which displays a friendly flash message
instead of the Rails error page.

Side effects of this change:

None.