Fix bug that prevents theses from being destroyed
Why these changes are being introduced:

The data model is set up in such a way that a Thesis with an associated
User cannot be destroyed. The desired behavior is that a Thesis can be
destroyed, and that will cascade to its associated Author entry without
destroying the User.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-182

How this addresses that need:

* Adds dependent: :destroy option to the has_many: :authors association
in the Thesis model.
* Adds dependent: :restrict_with_error option to User-Thesis assocation,
so users see a flash message instead of a Rails error page when they try
to destroy a User with an associated Thesis.
* Adds test coverage confirming desired behavior.

Side effects of this change:

None.