Update fields in Hold and Hold Source UIs
Why these changes are being introduced:

This is part of the work to implement recommendations made in the UX
review of the Hold and Hold Source UI. This changeset is detailed in
ETD-178, and originated in ETD-106 (see links in relevant tickets
section).

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-178
* https://mitlibraries.atlassian.net/browse/ETD-106
* https://mitlibraries.atlassian.net/browse/ETD-170

How this addresses that need:

* Adds degrees field to Hold show view
* Adds degree date (thesis.grad_date) field to Hold show and index views
* Removes holds field from Hold Source create and edit views
* Removes source field from Hold Source show view
* Adds expired_holds field to Hold Source index view

Side effects of this change:

It was requested that we remove the holds field from the Hold Source
create view, but administrate requires us to remove it from the edit
view as well. We should confirm with stakeholders that this is acceptable
in QA.