Update date format in Hold and Hold Source UIs
Why these changes are being introduced:

The current administrate field (DateTime) for created_at and updated_at
includes the timestamp, which adds visual clutter. This change was
requested as part of the UX review of the Hold UI.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-175
* https://mitlibraries.atlassian.net/browse/ETD-106
* https://mitlibraries.atlassian.net/browse/ETD-170

How this addresses that need:

This changes the DateTime fields in the Hold and Hold Source UIs to Date
Fields, thus removing timestamps.

Side effects of this change:

None.