Adds thesis author names to the Hold UI
Why these changes are being introduced:

Author names are crucial for processors to clarify the correct
Thesis-Hold association, as they are more consistent than thesis titles.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-107
* https://mitlibraries.atlassian.net/browse/ETD-177
* https://mitlibraries.atlassian.net/browse/ETD-106
* https://mitlibraries.atlassian.net/browse/ETD-170

How this addresses that need:

* Adds an `author_names` convenience method to the Hold model that maps
all of the Author names associated with the corresponding Thesis.
* Adds `author_names` field to the Hold admin dashboard show and index
views.

Side effects of this change:

None, but adding author info to the new and edit view will need to be
implemented separately (currently documented in ETD-177, linked above).