Confirm audit trail updates when moving a hold to another thesis
Why these changes are being introduced:

We want to confirm both that a hold can be moved to another thesis and
that the audit updates accordingly when such a move is made.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-152

How this addresses that need:

Moving holds is already tested in the Thesis model spec. This PR adds
testing that confirms the audit trail piece.

Side effects of this change:

* The ticket mentions that thesis processors should be able to do this in
the UI, but thesis processors can't access the admin dashboards at all.
I've created a ticket for that here: https://mitlibraries.atlassian.net/browse/ETD-190
* We currently move holds to a new thesis based on the thesis title.
Pending stakeholder feedback, we may want to change that.