Add administrate views for customization
Why these changes are being introduced:

In order to implement ETD-177, we need to customize some of the
default administrate views. We want to add the upstream views as a
separate commit so it is clear in the commit history what changes
were made.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-177

How this addresses that need:

* Adds upstream views for customization: Thesis show, Hold form partial.
* Adds upstream layout to include custom CSS.

Side effects of this change:

Additional overhead, as we will need to maintain these files if/when
they change in administrate.