Add logic to create a Hold through Thesis
Why these changes are being introduced:

When creating or editing a hold, processors need to be able to clearly
identify its parent record. The selectize drop-down displays the title,
but processors also need to be able to search by author name and kerb ID,
and the corresponding degree info must be visible.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-177

How this addresses that need:

* Adds a 'New hold' button to the Thesis show page that links to a new
Hold form prepopulated with information about the corresponding Thesis.
* Overrides the administrate HoldsController#new action to work with the
thesis_id param.
* Adds a search field to the Authors dashboard index page to search by
display_name, kerberos_id, and uid.
* Adds kerberos_id to Users dashboard index page to provide another
path to access the desired thesis.
* Adds a new stylesheet for the thesis info panel (anticipating that
we may need additional custom CSS for administrate in the future).

Side effects of this change:

* Additional maintenance overhead from customizing upstream view files
* After extensive internal discussions in EngX, we still don't have an
effective means of clarifying the associated thesis when editing a hold.
The best approach so far is to modify the display_resource method for
the thesis dashboard, which is not ideal as it affects other parts of
the UI. More info on this problem is available in the JIRA ticket.

Co-authored-by: Jeremy Prevost <jprevost@mit.edu>