Add relevant audit trail fields to Holds UI
Why these changes are being introduced:

Processors need certain information regarding a Hold's audit trail in the
admin UI:

* The date the hold was released
* The date the thesis file was received
* The user who created the hold

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-176

How this addresses that need:

* Adds new fields to the Hold dashboard: created_by,
date_thesis_file_received, date_released.
* Adds convenience methods to the Hold model to access the necessary
data.

Side effects of this change:

Per internal discussions, we chose to conflate the date the thesis record
was created with the date the thesis file was received. It's worth
noting here that these may not always be the same.