Implement requested UX changes for Hold UI
Why these changes are being introduced:

This addresses UX feedback given in response to the recent Hold UI updates
(mocked up in ETD-106 and implemented across various other tickets).

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-189

How this addresses that need:

* Adds an i18n translation to render processing_notes field as 'Hold
notes'.
* The display_resource method now returns "Hold for: #{thesis.title}" to
clarify that the current resource is a Hold.
* Renders dates_thesis_files_received in form as a read-only element.

Side effects of this change:

* If a new Hold is created without a thesis ID param, the form won't
populate the dates_thesis_files_received field dynamically based on the
contents of the thesis field. This shouldn't be a problem, since the
workflow will be to create holds through theses.
* Adding dates_thesis_files_received to the form broke the 'new' action
if no thesis ID param was provided. I updated the method to check for
an associated thesis before attempting to render the dates.