Create UI for full listing of hold history
Why these changes are being introduced:

We need a way to see the full audit trail for a Hold record. Only some
paper_trail fields are currently visible in the Hold administrate UI.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-186

How this addresses that need:

* Adds a hold history page (the Hold show page) that displays audit
information for a given Hold record. Initially, I chose to display the
following fields for each version:
    * The event type (e.g., create/update/destroy);
    * The user who modified the record; and
    * Everything in the version changeset.
* Adds a custom field, hold_history, to administrate, to provide a link
to the hold history page from the Hold dashboard.
* Adds Hold#show method to route to the hold history page.

Side effects of this change:

* Added complexity: we now have a Hold controller with a show method, a
corresponding view and route, and a Hold helper.
* In order to get the custom field to work and display as desired, I
added an i18n translate to render the ID field in the Hold dashboard
as 'Hold history'. We may need to rework this later, if for some reason
we want to display the Hold ID in a different context.
* Because I chose to create this view outside of administrate, it uses
the regular Thing layout. This might be confusing to users since they
will be coming to this view from the admin dashboard, which has a
different layout.