Add rake tasks to seed degrees and departments
Why these changes are being introduced:
We have several hundred official degrees and departments with unique
codes and names from the Data Warehouse that need to be seeded to the
database. Ideally this process will be repeatable for initial testing
and future dev/test/staging database resets.

How this addresses that need:
* Adds two db-namespaced rake tasks, one to seed the degrees and
  another to seed the departments.

Side effects of this change:
* In doing the initial full department seed, discovered that the data
  warehouse dept names are not unique. This meant the unique constraint
  on Department.name_dw had to be removed.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/etd-169