Implement public UI home page
Why these changes are being introduced:

We need a landing page for the public UI to direct students and
transfer submitters to the correct form.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-212
* https://mitlibraries.atlassian.net/browse/ETD-209 (mockup)

How this addresses that need:

This modifies the existing static index page to reflect the new app's
purpose and functionality.

Side effects of this change:

* Transfer submitters are now redirected after login to the transfer
submission form.
* Removes an extra newline from the PR template (completely unrelated,
but it was irking me).