Update transfer submission UI
Why these changes are being introduced:

The UX review of the transfer submission UI recommended some changes.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-196
* https://mitlibraries.atlassian.net/browse/ETD-80 (review and mockup)

How this addresses that need:

This updates the transfer submission UI to reflect the mockup in ETD-80.

Side effects of this change:

* Adds an attr_accessor method to the Transfer model to make the
certification boolean input work. Mikki has confirmed that this value
does not need to be saved to the db.
* Moves the 'you_are' partial to 'views/shared' (thesis views have been
updated accordingly).