Fix slim footer
Why these changes are being introduced:

The slim footer is broken due to an issue with the pre-release of the
mitlibraries-theme gem we are using in this app.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-222

How this addresses that need:

This updates the mitlibraries-theme gem to v0.5.0, which includes a fix
for the footer as well as Rails 6 support.

Side effects of this change:

Dependabot/depfu PRs will be opened in our Rails 5 apps to update
mitlibraries-theme. This shouldn't break anything in those apps, but we
should make sure there aren't any unintended effects before we merge
those PRs.