Update student submission mailer for new application
Why these changes are being introduced:

We want to send a helpful confirmation email to students when they
submit their thesis info.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-200

How this addresses that need:

This updates the ReceiptMailer to include the information specified in
in the contextual content doc. The mailer is called in the update action
of the Thesis controller.

Side effects of this change:

* Some weird spacing in the receipt_email fixture was required to pass
the assertion comparing the fixture to the email body.
* Emails are now sent _only_ if `DISABLE_ALL_EMAIL` is set to `false`.
Our previous guard clause disabled email if that ENV was set, so this
behavior is similar, but hopefully more intuitive.