Remove degree field from new thesis form
Why these changes are being introduced:

The degree drop-down is difficult to use due to duplicate entries. The
degree will always be supplied by the registrar, so we don’t need it from
creation.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-230

How this addresses that need:

This removes the degree validation from the thesis model and removes the
degree field from the new form.

Side effects of this change:

A confirmation email sent to a student that completes the new form will
show no degree associated with their thesis, which may be confusing.