Change degree year field to a select
Why these changes are being introduced:

The degree year field in the thesis form is currently free text, but
the stakeholders would prefer that it be a drop-down with last year,
current year, and next year as the options.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-250

How this addresses that need:

* Changes the field in thesis/new to a select using simpleform's collection
option, which allows us to specify the values in an array.
* Uses the `last_year` and `next_year` methods on the `Date` object to
evaluate which years to add to the collection.

Side effects of this change:

None.