Open thesis form outlinks in a new tab/window
Why these changes are being introduced:

Form links currently open in the same window, which could be frustrating
for users if they lose the data they've entered in the form.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-251

How this addresses that need:

This adds target="_blank" to anchor tags that link to external sites to
force browsers to open those links in a new tab/window.

Side effects of this change:

I railsified the link to the root path in the site_nav partial. This
doesn't affect anything, but it felt nice to do.