Open mailto links in a new window/tab
Why these changes are being introduced:

If a user has configured a browser-based email client as their default,
then mailto links will open in the same window. We want those links to
open in a new window. This is stakeholder feedback on ETD-251.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-251

How this addresses that need:

Adds target="_blank" to mailto links in the landing page and thesis
partials.

Side effects of this change:

None.