Adds json gem
Why are these changes being introduced:

* Most versions of ruby seem to include this gem bundled, but for some
reason this is not happening on my latest laptop. This experimental
branch is to mostly confirm adding this gem manually doesn't negatively
affect our various environments.

Document any side effects to this change:

We're explicitly including a gem that seems to often already be
included. This may allow us to keep it up to date more than relying
on ruby-included versions that may be installed in various environments.