Bugfix for broken Hold dashboard searches
Why these changes are being introduced:

Certain fields in the Hold dashboard, which are generated via convenience
methods and not stored in the db,  were defined as Field::String. Because
Field::String is searchable by default, this caused the dashboard search
to break when it tried to query nonexistent columns.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-304

How this addresses that need:

This changes the offending fields to Field::Text, which are not
searchable by default.

Side effects of this change:

* When I fixed this bug, I noticed that the only searchable field was
status. I added thesis title as another searchable field, since there
aren't many Hold model fields that are searchable in administrate.
* Notably, status is searchable only if you add a colon to the
end of the querystring (e.g., 'active:'), due to some weirdness with
how collection filters work in administrate.