Fixes a typo in the import job
Why these changes are being introduced:
A typo in the import job caused the job to crash when an expected error
was logged. This fixes it.

How this addresses that need:
* Fixes the typo

Side effects of this change:
Note: because this is semi-urgent, this commit does not update tests to ensure that this particular rescue block is covered. I will add a Jira ticket for that to make sure the tests are updated ASAP.