Display additional data in hold history view
Why these changes are being introduced:

Certain data associated with a hold (author names, degree information,
and dates thesis files received) are not included in the hold history
page because these fields are not part of the Hold model and are
therefore not under audit control.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-341

How this addresses that need:

This adds the current values of the fields mentioned to the top of the
hold history for users to reference while reviewing a hold's audit
trail.

Side effects of this change:

I refactored a couple of conditionals into guard clauses in the Hold model.