Make author names searchable on Hold dashboard
Why these changes are being introduced:

Author names were not searchable in the Hold dashboard because
administrate accesses them from a convenience method rather than the
User model, causing the the Active Record query to fail.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-353

How this addresses that need:

This adds an association to the User model on the Hold model, and a
corresponding 'users' attribute to the Hold dashboard, which is searchable
on display_name, uid, and kerberos_id. These are the same fields that are
currently searchable on the Author dashboard, but we can add more fields
if it proves useful.

Side effects of this change:

None, aside from the new association. We are still using the author_names
convenience method to display the names in the UI, since the has_many
association renders in a not-so-useful way in administrate.