Add papertrail to Thesis and User models
Why these changes are being introduced:

Audit trails are needed for the User and Thesis models to enhance
registrar data import reporting.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-357

How this addresses that need:

This adds papertrail to Thesis and User models.

Side effects of this change:

As noted in the ticket, we may later want to expand papertrail to all
(or most) of the other models. For now, we are limiting scope to the
two models that directly affect the registrar data import.