Generate email report for data registrar import jobs
Why these changes are being introduced:

Processors and thesis admins need to see summary data about the
registrar import job, especially if there are any issues with it.
The job currently logs much of this information, but an email report
would be more useful for stakeholders to review.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-354
* https://mitlibraries.atlassian.net/browse/ETD-348
* https://mitlibraries.atlassian.net/browse/ETD-371

How this addresses that need:

This adds some data collection to the import job, as well as a mailer
that reports the following data:

* Number of theses in the job
* Number of theses processed
* Errors found (total number and list)
* New DLCs added (total number and list)
* New degrees added (total number and list)
* Number of new users
* Number of new theses
* Number of updated theses

Side effects of this change:

* Small config change to default_url in the production environment
to accommodate urls in emails. We are also now using PREFERRED_DOMAIN
to generate links in the email for staging and prod.
* Added letter_opener gem to make emails easier to review locally.