Reword language in registrar import email
Why these changes are being introduced:

Stakeholder feedback on this email was to changing the wording from
'total theses' to 'total rows', as each row is not necessarily a
distinct thesis.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-354

How this addresses that need:

This makes the requested changes.

Side effects of this change:

None