Change creative commons link
Why these changes are being introduced:

This was additional stakeholder feedback on the student submission
form.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-390

How this addresses that need:

This updates the creative commons link on the student submission
forms.

Side effects of this change:

None at first, although the link is to a beta version of the site,
so we may need to update it in the future.