Update dependencies
Why these changes are being introduced:

Updates are available for bootsnap, aws-sdk-s3, puma, rubocop,
and listen. We are not yet able to update devise and omniauth, as
we will need to change how we handle logins to accommodate the
updates to those gems.

Relevant ticket(s):

N/A.

How this addresses that need:

This updates the aforementioned dependencies.

Side effects of this change:

None identified during initial UI testing. Thus far I've checked
the transfer submission and processing workflows, thesis submission
and processing workflows, and registrar data import.