Add processor note to searchable fields in thesis processing queue
Why these changes are being introduced:

This was a requested change from the first round of thesis processing.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-352

How this addresses that need:

This makes processor_note a searchable field in the thesis processing
datatable. We are leaving the column visible for now, but we may
need hide it later if it creates too much clutter in the UI. Jess will
keep an eye on this while processing theses.

Side effects of this change:

This is the only text field that we are querying in the datatable,
so it's an open question as to whether this will perform adequately
at scale. We can adjust as needed based on what we see in prod.