Construct DSS JSON
Why these changes are being introduced:

We want to be able to publish theses to DSpace, which requires
us to convert relevant thesis metadata to DSS-compliant JSON.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-394

How this addresses that need:

This creates the requisite submission metadata for DSS, based on the
submission message spec and the data mappings spreadsheet. It does not
save said metadata to AWS, nor does it write messages to the queue.

Side effects of this change:

There are fields that are required for publication that are not
validated anywhere in the application. The current process is for
Jess to confirm these as part of their workflow, but we should make
their life easier by including an automated check as part of
the publication status update workflow. A ticket for this work is
open here: https://mitlibraries.atlassian.net/browse/ETD-431

Co-Authored-By: Jeremy Prevost <jprevost@mit.edu>