Upload DSS metadata JSON files to S3
Why these changes are being introduced:

This is additional work toward ETD-394, which involves creating
a DSS-compliant JSON package and submitting a message to the SQS
queue. This particular commit associates metadata JSON objects
with Thesis records, and puts the metadata files in place in S3.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-394

How this addresses that need:

This uses Active Storage to attach metadata JSON (i.e., DspaceMetadata
objects) to their Thesis records and upload them to S3. The next
step is to build the SQS message and the job that puts it all together
and submits it to the SQS queue.

Side effects of this change:

We're using StringIO to simulate IO on the serialized JSON string, so
Active Storage will attach it as a file. This doesn't affect the
outcome, but it's worth noting as a fairly unusual approach.

Remove publication job shell