Construct DSS SQS message
Why these changes are being introduced:

This is the third part of ETD-394, supporting publication to
DSpace@MIT via the DSpace Submission Service.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-394

How this addresses that need:

This constructs the SQS message attributes and body, as per the
DSS spec. Submitting the message to the queue will come in a
subsequent PR.

Side effects of this change:

None.