Send messages to SQS queue for DSpace publication
Why these changes are being introduced:

In order to publish theses to DSpace, we need to send messages to
an SQS queue that the DSpace Submission Service (DSS) can read from
and continue the publication workflow.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-394

How this addresses that need:

This creates a DspacePublicationJob that constructs and sends
a DSS-compliant message to the SQS queue.

Side effects of this change:

* The AWS Ruby SDK requires message attribute keys to be snake case
instead of camel case and symbols instead of strings. It's unclear
if this will cause issues with DSS.
* Added a message to development.rb about the ActiveStorage host.
Currently this needs to be each time one starts a rails console,
but there is an open PR in the rails codebase that will allow us
to configure it for the dev environment.
* We chose to use AWS client stubs for testing purposes rather than
VCR. This means that we'll be instantiating the SQS client outside
of the job, but we don't anticipate any issues with this approach.

co-authored-by: Jeremy Prevost <jprevost@mit.edu>