Updates metadata to match DSS expectations
Why these changes are being introduced:

The DSpace Submission Queue expects us to supply duplicate key names
rather than an array of values for any repeated metadata values.

Relevant ticket(s):

* https://mitlibraries.atlassian.net/browse/ETD-394

How this addresses that need:

To create a Hash in ruby with duplicate key names, which is useful so
we can easily use `.to_json`, we need to use `.compare_by_identity`
when building the Hash.

https://ruby-doc.org/core-2.7.1/Hash.html#method-i-compare_by_identity

We also extend the expiration time for the S3 URLs we include to the
maximum value of one week so they don't expire before DSS finishes
processing the queue.

Side effects of this change:

`.compare_by_identity` does not allow you to access non-symbol values
in a hash as each key is now comparing at the Object level and not at
the string level. Using symbols would solve that, but also would prevent
duplicate keynames again so we can't do that. For this application, this
is not a problem but it did require a different approach to testing as
we could not just look at the hash and instead needed to look at the
final output to see if it has the data we expect. This is minor, but the
tests are a bit more verbose as a result. The test cover the same
functionality. This is annoying, but the end result is both data in the
expected format for DSS _and_ code being tested.