Read from SQS output queue, update pub statuses and handles
Why these changes are being introduced:

Once a thesis is published (or fails to publish) to DSpace, we
want to update the corresponding record in ETD with its DSpace
handle and the appropriate publication status.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/ETD-399

How this addresses that need:

This creates a DspacePublicationResults job that reads from the
SQS output queue, attempts to match messages to records in ETD,
and updates the ETD records with the DSpace handle (if available)
and publication status (either 'Published' if successful or
'Publication error') if not. Once the job is complete, an email report
is sent if any theses were processed.

Side effects of this change:

* Adds 'Publication error' to the list of valid pub statuses on the Thesis
model.
* Updates the input/output queue names and URLs to match the sample data
we use in local development.
* The email report may need further refinement, pending stakeholder
feedback.
* The Aws::SQS::QueuePoller.poll method does not seem to work with the
latest version of moto, which we use for local development. We currently
have a branch of DSS with a pinned version of moto as a temporary fix.
* The poll_messages method has nested loops. We are accepting this
complexity for now, as moving the nested loop to its own method caused
other issues.

co-authored-by: Jeremy Prevost <JPrevost@users.noreply.github.com